attribute test auth

+++ tests/test_attribute.py
@@ -1,15 +1,16 @@
 from fastapi.testclient import TestClient
 from fastapi import status
 
-def test_read_attribute(client: TestClient):
+def test_read_attribute(client: TestClient, admin_token_headers: dict):
     client.post(
         "/attribute/",
         json={
             "title": "Color",
         },
+        headers=admin_token_headers,
     )
 
-    response = client.get("/attribute/?limit=1")
+    response = client.get("/attribute/?limit=1", headers=admin_token_headers)
     data = response.json()
 
     assert data["count"] == 1      
@@ -20,28 +21,30 @@ def test_read_attribute(client: TestClient):
     assert isinstance(result["id"], int)
     assert result["title"] == "Color"
 
-def test_read_attribute_by_id(client: TestClient):
+def test_read_attribute_by_id(client: TestClient, admin_token_headers: dict):
     create_response = client.post(
         "/attribute/",
         json={
             "title": "Color",
         },
+        headers=admin_token_headers,
     )
     attribute_id = create_response.json()["id"]
 
-    response = client.get(f"/attribute/{attribute_id}")
+    response = client.get(f"/attribute/{attribute_id}", headers=admin_token_headers)
     
     assert response.status_code == status.HTTP_200_OK
     result = response.json()
     assert result["id"] == attribute_id
     assert result["title"] == "Color"
 
-def test_create_attribute(client: TestClient):
+def test_create_attribute(client: TestClient, admin_token_headers: dict):
     response = client.post(
         "/attribute/",
         json={
             "title": "Color",
         },
+        headers=admin_token_headers,
     )
 
     assert response.status_code == status.HTTP_200_OK
@@ -49,12 +52,13 @@ def test_create_attribute(client: TestClient):
     assert isinstance(result["id"], int)
     assert result["title"] == "Color"
 
-def test_update_attribute(client: TestClient):
+def test_update_attribute(client: TestClient, admin_token_headers: dict):
     create_response = client.post(
         "/attribute/",
         json={
             "title": "Color",
         },
+        headers=admin_token_headers,
     )
     attribute_id = create_response.json()["id"]
 
@@ -63,6 +67,7 @@ def test_update_attribute(client: TestClient):
         json={
             "title": "Color 2",
         },
+        headers=admin_token_headers,
     )
 
     assert response.status_code == status.HTTP_200_OK
@@ -70,14 +75,15 @@ def test_update_attribute(client: TestClient):
     assert isinstance(result["id"], int)
     assert result["title"] == "Color 2"
 
-def test_delete_attribute(client: TestClient):
+def test_delete_attribute(client: TestClient, admin_token_headers: dict):
     create_response = client.post(
         "/attribute/",
         json={
             "title": "Color",
         },
+        headers=admin_token_headers,
     )
     attribute_id = create_response.json()["id"]
 
-    response = client.delete(f"/attribute/{attribute_id}")
+    response = client.delete(f"/attribute/{attribute_id}", headers=admin_token_headers)
     assert response.status_code == status.HTTP_200_OK
